# swagger.yaml
openapi: 3.0.0
info:
  title: Turkish Recipe API
  version: 1.0.0
  description: Simple REST API for Turkish recipes

servers:
  - url: /
    description: API Server

paths:
  /api/recipes:
    get:
      summary: Get all recipes
      parameters:
        - in: query
          name: categoryId
          schema:
            type: string
          description: Filter by category ID
      responses:
        200:
          description: List of recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeListItem'

  /api/recipes/{id}:
    get:
      summary: Get recipe by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Recipe ID
      responses:
        200:
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        404:
          description: Recipe not found

  /api/categories:
    get:
      summary: Get all categories
      responses:
        200:
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  /api/categories/{id}:
    get:
      summary: Get category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Category ID
      responses:
        200:
          description: Category details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
        404:
          description: Category not found

components:
  schemas:
    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: string
        instructions:
          type: array
          items:
            type: string
        categoryId:
          type: string
        nutrition:
          type: object
          properties:
            calories:
              type: number
            protein:
              type: number
            carbohydrates:
              type: number
            fat:
              type: number

    RecipeListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string